fix duplicate dashboard fetches
